requirements.txt
asgiref==3.8.1
click==8.1.7
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.30.6
whitenoise==6.7.0
